[Bugfix] Use four instead of one ResourceSet in performance benchmark
In the unfixed version, the performance benchmark loads each of the four pair of metamodels into the same resource set (within the "getModel()" method). This appears to violate a global assumption of EMF: ResourceSet.getPackageRegistry().getEPackage(nsURI) returns a single package, and not a list of packages. Since my code relies on this global assumption, I submit a fix, which ensures that each pair of metamodels (together with the instances) is loaded into a separate ResourceSet.